will output cf template in json – bumping node version
