--- .github/workflows/ci.yml
@@ -1,12 +1,0 @@
-on:
-  workflow_dispatch:
-
-jobs:
-  trigger-circleci:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Run tests
-        id: tests
-        uses: CircleCI-Public/trigger-circleci-pipeline-action@v1.0.5
-        env:
-          CCI_TOKEN: ${{ secrets.CCI_TOKEN }}
