+++ index.test.ts
@@ -1,0 +1,8 @@
+import {Application} from './src/index';
+
+const test = new Application({
+    domainName: 'aaronwest.me',
+    webAssetPath: 'dist',
+});
+
+console.log(test);
