+++ package.json
@@ -1,9 +1,13 @@
 {
   "name": "@elevator-robot/blip",
-  "version": "1.0.31",
+  "version": "1.0.54",
   "description": "Build a simple static site",
   "main": "dist/index.js",
-  "type": "module",
+  "types": "dist/index.d.ts",
+  "type": "commonjs",
+  "files": [
+    "dist"
+  ],
   "scripts": {
     "lint": "eslint .",
     "test": "tsc --build tsconfig.json && jest --runInBand",
@@ -14,9 +18,6 @@
     "type": "git",
     "url": "https://github.com/elevator-robot/blip"
   },
-  "files": [
-    "dist"
-  ],
   "keywords": [],
   "author": {
     "name": "Aphexlog",
@@ -36,12 +37,12 @@
   "bugs": {
     "url": "https://github.com/elevator-robot/blip/issues"
   },
-  "types": "src/index.d.ts",
   "homepage": "https://github.com/elevator-robot/blip#readme",
   "devDependencies": {
     "@types/node": "^18.11.18",
     "aws-cdk-lib": "^2.58.1",
     "constructs": "10.1.203",
+    "ts-node": "^10.9.1",
     "typescript": "^4.9.4"
   },
   "peerDependencies": {
