+++ tsconfig.json
@@ -9,7 +9,7 @@
     "experimentalDecorators": true,
     "incremental": true,
     "lib": [
-      "ES2020"
+      "ES2022"
     ],
     "module": "CommonJS",
     "newLine": "lf",
@@ -26,14 +26,15 @@
     "strictNullChecks": true,
     "strictPropertyInitialization": true,
     "stripInternal": false,
-    "target": "ES2020",
+    "target": "ES2022",
     "composite": false,
     "tsBuildInfoFile": "tsconfig.tsbuildinfo"
   },
   "include": [
-    "**/*.ts"
+    "**/*.ts",
   ],
   "exclude": [
-    "node_modules"
-  ]
+    "node_modules",
+    "dist",
+  ],
 }
